backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG Chatbot backend for Docusaurus textbook"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "qdrant-client>=1.12.0",
    "openai>=1.50.0",
    "openai-agents>=0.0.3",
    "tiktoken>=0.8.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
